+++ src/Ext/Aircraft/Hooks.cpp
@@ -1258,3 +1258,13 @@ DEFINE_HOOK(0x418782, AircraftClass_CurleyShuffle_Default, 0x6)
 }
 
 #pragma endregion
+
+DEFINE_HOOK(0x415A00, AircraftClass_Mission_Paradrop_Overfly_Delay, 0x5)
+{
+	GET(AircraftClass*, pThis, ESI);
+
+	const auto pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->Type);
+	R->EAX(pTypeExt->ParadropDelay.Get(RulesExt::Global()->ParadropDelay));
+
+	return 0x415A05;
+}