+++ src/Ext/TechnoType/Body.h
@@ -488,6 +488,7 @@ public:
 
 		Nullable<Mission> ParadropMission;
 		Nullable<Mission> AIParadropMission;
+		Nullable<int> ParadropDelay;
 
 		Nullable<int> PenetratesTransport_Level;
 		Valueable<double> PenetratesTransport_PassThroughMultiplier;
@@ -963,6 +964,7 @@ public:
 
 			, ParadropMission {}
 			, AIParadropMission {}
+			, ParadropDelay {}
 
 			, PenetratesTransport_Level {}
 			, PenetratesTransport_PassThroughMultiplier { 1.0 }